Added a ME field to use the encoder to monitor the position when a motor has in/external encoder, but the encoder is disbled (EE=0) so that backlash can be properly handled.
In an ideal situation when a motor has enough driving power, it will
not miss steps, the positions calculated from the motor steps and from
encoder counts should always agree.  If these two disagree (MISS=1),
it means the motor is missing steps, and needs more running current.
Or, we cannot rely on it to do backlash correction.


git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/IMS/current@20979 ececa60f-9f63-0410-97c7-ea05888eb260